refactor: expose connected_components as free function, make UnionFind module-private
